Fix Terminal::keyPress() for non-ASCII chars (Fixes JB#28668)
On some installations (e.g. tablet), Terminal::keyPress() messes
up the keycode by converting the keypress into a Latin1 character
and then appending this Latin1 character to a QString. For some
reason, we were lucky so far and it worked on the phone.

Behavior change: This now causes key combinations with Ctrl to be
ignored if the key isn't in the ASCII ranges 0x41-0x5f, 0x61-0x7f.

- thp